app/requirements/requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
sqlalchemy==2.0.44
asyncpg==0.30.0
pydantic==2.12.4
pydantic-settings==2.11.0
pandas==2.3.3
python-telegram-bot==22.5
python-multipart==0.0.20
aiohttp==3.13.2
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
